+++ examples/my-app/CHANGELOG.md
@@ -1,5 +1,27 @@
 # my-app
 
+## 1.47.0
+
+### Minor Changes
+
+- fix bugs
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/provider@1.48.0
+
+## 1.46.0
+
+### Minor Changes
+
+- fix xax gas price multiplier
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/provider@1.47.0
+
 ## 1.45.0
 
 ### Minor Changes