+++ packages/wallet/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @0xsodium/wallet
 
+## 1.48.0
+
+### Minor Changes
+
+- fix bugs
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/abi@1.48.0
+  - @0xsodium/config@1.48.0
+  - @0xsodium/network@1.48.0
+  - @0xsodium/transactions@1.48.0
+  - @0xsodium/utils@1.48.0
+
+## 1.47.0
+
+### Minor Changes
+
+- fix xax gas price multiplier
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/network@1.47.0
+  - @0xsodium/abi@1.47.0
+  - @0xsodium/config@1.47.0
+  - @0xsodium/transactions@1.47.0
+  - @0xsodium/utils@1.47.0
+
 ## 1.46.0
 
 ### Minor Changes